requirements.txt
# FastAPI service
fastapi>=0.115,<0.116
uvicorn[standard]>=0.27,<0.31
python-multipart>=0.0.9
pydantic>=2.6,<3

# Document conversion (MarkItDown handles Office files)
markitdown[all]>=0.0.1    # Office documents, text files (includes docx, pptx, xlsx extras)
openpyxl>=3.1.0           # xlsx merged-cell preprocessing (also transitive via markitdown)
Pillow>=10.0.0            # Image handling for MarkItDown

# Email parsing
extract-msg>=0.48.0        # Parse Outlook .msg files

# Helpers
filetype>=1.2.0

# PDF extraction (PyMuPDF)
pymupdf>=1.24.0

# Document generation
weasyprint>=60.0
python-docx>=1.1.0
markdown>=3.5.0

# Docling proxy HTTP client
httpx>=0.27.0
